requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncmy
pydantic-settings
jinja2
python-multipart
